Add files via upload
Project Architecture (Source :Udemy: Azure Data Factory For Data Engineers - Project on Covid19 by Ramesh Retnasamy. 

We'll ingest the COVID-19 data from the European Center for Disease Prevention and Control website. Otherwise called ECDC. We use the HTTP connector
within Azure Data Factory to get the data. For the Eurostat data,we'll keep the population file in an Azure storage account and ingest from there. We could get this data directly from the Eurostat website, but we'll instead ingest from an Azure storage account.  Both data sets will be ingested into the Azure Data Lake storage.

We'll then use Azure Data Factory to transform this data. We'll use the transformation capability called Data Flows within the Data Factory for the couple of data sets.
We'll use HDInsight for another data set and Azure data bricks for another one. For the HDInsight and the data bricks transformations, Data Factory will be used as the orchestration tool rather than the transformation tool. All of the transformed data will then be stored in our Data Lake for use in machine learning models. We won't be building the machine learning models since we're only focusing on Data Factory purely..

We'll also push the subset of the required datato the SQL database so that it can be used for reporting.

We'll not build a Power BI report from this data. The focus of this course is Azure Data Factory rather than Power BI. But I'll do a quick lesson on building the report.
Let's now have a look at the solution to understand why we've used these specific technologies. Firstly, we've used Azure Data Factory for all our data integration and orchestration. We used ADF for data integration because it provides connectors to all our data sources in this project. And also it provides a vast array of connectors that could be used to expand the project in the future when a new requirement arises.

Similarly, Data Factory has the connectors to orchestrate our workflow. For example, it can orchestrate the workflow with running transformations in HDInsight and Azure databricks. Next, let's have a look at the transformation technologies being used. We've used three transformation technologies here, Data Flow within Data Factory, HDInsight, and Azure Databricks.

The storage solutions used: We're using Azure Blob storage to keep the population data, Azure Data Lake storage Gen 2 as the Data Lake, Azure SQL database for our data warehouse solution. We've used as Azure Data Lake Gen 2 for our Data Lake.
Azure Data Lake Gen 2. And finally, we've used Azure's SQL database for our data warehouse solution. Azure's synapse analytics formally called Azure data warehouse, is a better solution for large data warehouses due to its massively parallel processing architecture. We're using SQL database instead. Mainly due to the small amount of data we are processing.